add bytecode code implementation
This commit adds a implementation of compiler.Code in python bytecode as mentioned in issue #25.
It is implemented using a generator and currently slower than the generic implementation to
allow for variables to be passed to .execute.
If MAKE_FUNCTION gets faster or the compiler subsystem gets changed it may be faster.