enabled the new caching subsystem
This commit enables the caching system mentioned in issue #23.
To allow for the new more complex configuration and more native types,
the config format has been changed from INI to TOML.
Furthermore, the config file is now searched in multiple places and a RuntimeError
is raised on failure.
The old cache handlers are removed when they are implemented using the new system.